requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
cssselect==1.1.0
discord.py==1.7.2
finviz @ git+https://github.com/mariostoev/finviz@751ea997a490afbb0a88da8a5c021d41cd2147de
Flask==2.0.1
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
lxml==4.6.3
MarkupSafe==2.0.1
multidict==5.1.0
python-dotenv==0.17.1
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
tenacity==7.0.0
tqdm==4.61.0
typing-extensions==3.10.0.0
urllib3==1.26.5
user-agent==0.1.9
Werkzeug==2.0.1
yarl==1.6.3
